Cleanup legacy utils and plotting functions
Remove unneeded functions, add notes about what is needed
in ApbdQuery, refactor legacy plotting utilities where feasible,
and add fully ticketed TODOs to all remaining legacy functions.
