+++ python/lsst/analysis/ap/legacyPlotUtils.py
@@ -21,37 +21,187 @@
 
 import numpy as np
 import matplotlib as mpl
-import matplotlib.cm as cm
 import matplotlib.pyplot as plt
 import matplotlib.ticker as plticker
 import matplotlib.patches as patches
 import pandas as pd
+from deprecated.sphinx import deprecated
 
 from astropy import units as u
-from astropy import visualization as aviz
+from astropy.visualization import ZScaleInterval, SqrtStretch, ImageNormalize
 
 import lsst.daf.butler as dafButler
 import lsst.geom
 from lsst.ap.association import UnpackApdbFlags, TransformDiaSourceCatalogConfig
 import lsst.afw.cameraGeom as cameraGeom
 from lsst.obs.decam import DarkEnergyCamera
+from lsst.pipe.base import Struct
 
-from .legacyApdbUtils import loadExposures
+'''All of the functions herein are deprecated legacy functions.
+In the comments, each has an associated ticket, and once those tickets
+are complete so the plotting functionality is elsewhere, the legacy code
+should be removed.
+'''
 
 
-def plotDiaObjectHistogram(objTable, objFiltered,
-                           label1='All Objects',
-                           label2='Filtered Objects',
-                           title=''):
-    """Create a histogram showing how many DIA Sources
-    comprise the DIA Objects.
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def getPatchConstituents(repo, band='g', printConstituents=False, verbose=False,
+                         tractIndex=0, instrument='DECam', collections=[],
+                         skymapName='', coaddName='deepCoadd'):
+    """Learn which patches have data in them, and what dataIds (i.e.,
+    visits/exposures and detectors) comprise each patch.
 
     Parameters
     ----------
-    objTable : `pandas.DataFrame`
-        DIA Object Table.
-    objFiltered : `pandas.core.frame.DataFrame`
-        DIA Object Table that is a filtered subset of objTable.
+    repo : `str`
+        The path to the data repository.
+    band : `str`, optional
+        The filter name of the data to select from the repository.
+    printConstituents : `bool`, optional
+        Select whether to print detailed information on each
+            patch that has data.
+    verbose : `bool`, optional
+        Select whether to print all the patches with and without data.
+    tractIndex : `int`, optional
+        Tract index, default 0
+    instrument : `str`, optional
+        Default is 'DECam', used with gen3 butler only
+    collections : `list` or `str`, optional
+        Must be provided for gen3 to load the camera properly
+    skymapName : `str`, optional
+        Must be provided for gen3 to load the skymap. e.g., 'hsc_rings_v1'
+    coaddName : `str`, optional
+        Type of coadd, default `deepCoadd` (you might want `goodSeeingCoadd`)
+
+    Returns
+    -------
+    dataPatchList : `list` of `str`
+        The patches containing coadds in the repository.
+    constituentList : `list` of `numpy.ndarray`
+        The visit numbers of the calexps that contributed to each patch.
+    constituentCountList : `list` of `int`
+        The number of visits that contributed to each patch.
+    """
+    if not collections:
+        raise ValueError('One or more collections is required in gen3.')
+    if not skymapName:
+        raise ValueError('A skymap is required.')
+    if not instrument:
+        raise ValueError('An instrument is required.')
+    butler = dafButler.Butler(repo, collections=collections)
+    skymap = butler.get('skyMap', instrument=instrument, collections='skymaps', skymap=skymapName)
+    tract = skymap.generateTract(tractIndex)
+    everyPatchList = [tract.getSequentialPatchIndex(tract[idx]) for idx in range(len(tract))]
+    dataPatchList = []
+    constituentList = []
+    constituentCountList = []
+    for patch in everyPatchList:
+        try:
+            dataIdPatch = {'band': band, 'tract': tractIndex,
+                           'patch': int(patch), 'instrument': instrument,
+                           'skymap': skymapName}
+            coadd_test = butler.get(coaddName, dataId=dataIdPatch)
+        except (LookupError):
+            if verbose:
+                print('No data in this patch', patch)
+            continue
+        else:
+            if verbose:
+                print('This patch has a coadd', patch)
+            constituent = coadd_test.getInfo().getCoaddInputs().ccds['visit']
+            constituentCount = len(constituent)
+            if printConstituents:
+                print(patch, constituentCount)
+            dataPatchList.append(patch)
+            constituentList.append(constituent)
+            constituentCountList.append(constituentCount)
+    return dataPatchList, constituentList, constituentCountList
+
+
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def bboxToRaDec(bbox, wcs):
+    """Get the corners of a BBox and convert them to lists of RA and Dec.
+
+    Parameters
+    ----------
+    bbox : `lsst.geom.Box`
+        The bounding box to determine coordinates for.
+    wcs : `lsst.afw.geom.SkyWcs`
+        The WCS to use to convert pixel to sky coordinates.
+
+    Returns
+    -------
+    ra, dec : `tuple` of `float`
+        The Right Ascension and Declination of the corners of the BBox.
+    """
+    skyCorners = [wcs.pixelToSky(pixPos.x, pixPos.y) for pixPos in bbox.getCorners()]
+    ra = [corner.getRa().asDegrees() for corner in skyCorners]
+    dec = [corner.getDec().asDegrees() for corner in skyCorners]
+    return ra, dec
+
+
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def getRaDecMinMaxPatchList(patchList, tractInfo, pad=0.0, nDecimals=4, raMin=360.0, raMax=0.0,
+                            decMin=90.0, decMax=-90.0):
+    """Find the max and min RA and DEC (deg) boundaries
+       encompassed in the patchList
+    Parameters
+    ----------
+    patchList : `list` of `str`
+       List of patch IDs.
+    tractInfo : `lsst.skymap.tractInfo.ExplicitTractInfo`
+       Tract information associated with the patches in patchList
+    pad : `float`
+       Pad the boundary by pad degrees
+    nDecimals : `int`
+       Round coordinates to this number of decimal places
+    raMin, raMax : `float`
+       Initiate minimum[maximum] RA determination at raMin[raMax] (deg)
+    decMin, decMax : `float`
+       Initiate minimum[maximum] DEC determination at decMin[decMax] (deg)
+
+    Returns
+    -------
+    `lsst.pipe.base.Struct`
+       Contains the ra and dec min and max values for the patchList provided
+    """
+    for ip, patch in enumerate(tractInfo):
+        patchWithData = tractInfo.getSequentialPatchIndex(patch)
+        if patchWithData in patchList:
+            raPatch, decPatch = bboxToRaDec(patch.getOuterBBox(), tractInfo.getWcs())
+            raMin = min(np.round(min(raPatch) - pad, nDecimals), raMin)
+            raMax = max(np.round(max(raPatch) + pad, nDecimals), raMax)
+            decMin = min(np.round(min(decPatch) - pad, nDecimals), decMin)
+            decMax = max(np.round(max(decPatch) + pad, nDecimals), decMax)
+    return Struct(
+        raMin=raMin,
+        raMax=raMax,
+        decMin=decMin,
+        decMax=decMax,
+    )
+
+
+# TODO: DM-43095, to analysis_tools following DM-41531 (diaSource flags)
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def plotDiaObjectHistogram(objects, row_mask,
+                           label1='All Objects',
+                           label2='Filtered Objects',
+                           title=''):
+    """Create a histogram showing how many DIA Sources comprise the DiaObjects.
+
+    Parameters
+    ----------
+    objects : `pandas.DataFrame`
+        DiaObject Table.
+    row_mask : `array` [`bool`]
+        Mask array for objects to plot filtered objects.
     label1 : `str`
         Legend label for the first DIA Object Table.
     label2 : `str`
@@ -69,11 +219,11 @@ def plotDiaObjectHistogram(objTable, objFiltered,
     plt.ylabel('Object count', size=16)
     plt.ylim(0.7, 1e5)
     plt.yscale('log')
-    binMax = np.max(objTable['nDiaSources'].values)
-    plt.hist(objTable['nDiaSources'].values, bins=np.arange(0, binMax),
-             color='#2979C1', label=label1)
-    plt.hist(objFiltered['nDiaSources'].values, bins=np.arange(0, binMax),
-             color='#Bee7F5', label=label2)
+    binMax = np.max(objects['nDiaSources'].values)
+    plt.hist(objects['nDiaSources'].values, bins=np.arange(0, binMax),
+             color='#2979c1', label=label1)
+    plt.hist(objects.loc[row_mask, 'nDiaSources'].values, bins=np.arange(0, binMax),
+             color='#bee7F5', label=label2)
     plt.legend(frameon=False, fontsize=16)
     plt.gca().spines['top'].set_visible(False)
     plt.gca().spines['right'].set_visible(False)
@@ -81,20 +231,20 @@ def plotDiaObjectHistogram(objTable, objFiltered,
     return fig
 
 
-def plotDiaObjectsOnSky(rerunName, objTable, objFilter, hits):
+# TODO: delete this as part of DM-43201
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def plotDiaObjectsOnSky(title, objects, row_mask):
     """Create a plot of filtered DIAObjects on the sky.
 
     Parameters
     ----------
-    rerunName : `str`
-        Name of the directory at the end of the repo path.
-    objTable : `pandas.DataFrame`
+    title : `str`
+        Title to give to the plots.
+    objects : `pandas.DataFrame`
         DIA Object Table.
-    objFilter : `pandas.Series` of `bool`
-        Filter applied to create a subset (e.g., quality cut) from objTable.
-    hits : `boolean`
-        True for two panels with custom axes for the hits2015 dataset
-        False for a single plot with automatic axis limits
+    row_mask : `pandas.Series` of `bool`
+        Filter applied to create a subset (e.g., quality cut) from objects.
 
     Returns
     -------
@@ -104,73 +254,29 @@ def plotDiaObjectsOnSky(rerunName, objTable, objFilter, hits):
     """
     fig = plt.figure(facecolor='white', figsize=(10, 8))
 
-    if hits:  # two subplots
-        dec_set1 = (objTable['dec'] > -2) & objFilter
-        dec_set2 = (~dec_set1) & objFilter
-        plt.subplots_adjust(wspace=0.1, hspace=0)
-
-        # Panel 1: one HiTS field, on the right
-        ax1 = plt.subplot2grid((100, 100), (0, 55), rowspan=90, colspan=45)
-        plot1 = ax1.scatter(objTable.loc[dec_set1, 'ra'], objTable.loc[dec_set1, 'dec'],
-                            marker='.', lw=0, s=objTable.loc[dec_set1, 'nDiaSources']*8,
-                            c=objTable.loc[dec_set1, 'nDiaSources'], alpha=0.7,
-                            cmap='viridis', linewidth=0.5, edgecolor='k')
-        plt.xlabel('RA (deg)', size=16)
-        plt.ylabel('Dec (deg)', size=16)
-        ax1.spines['top'].set_visible(False)
-        ax1.spines['left'].set_visible(False)
-        ax1.yaxis.tick_right()
-        ax1.yaxis.set_label_position('right')
-        ax1.invert_xaxis()  # RA should increase to the left
-        loc = plticker.MultipleLocator(base=0.5)  # puts ticks at regular intervals
-        ax1.yaxis.set_major_locator(loc)
-        cb = fig.colorbar(plot1, orientation='horizontal', pad=0.3)
-        cb.set_label('Number of Sources per Object', size=16)
-        cb.set_clim(np.min(objTable.loc[dec_set2, 'nDiaSources']),
-                    np.max(objTable.loc[dec_set2, 'nDiaSources']))
-        cb.solids.set_edgecolor("face")
-        cb.remove()  # don't show colorbar by this panel
-
-        # Panel 2: two (overlapping) HiTS fields, on the left
-        ax2 = plt.subplot2grid((100, 100), (0, 0), rowspan=90, colspan=50)
-        plot2 = ax2.scatter(objTable.loc[dec_set2, 'ra'], objTable.loc[dec_set2, 'dec'],
-                            marker='.', lw=0, s=objTable.loc[dec_set2, 'nDiaSources']*8,
-                            c=objTable.loc[dec_set2, 'nDiaSources'], alpha=0.7,
-                            cmap='viridis', linewidth=0.5, edgecolor='k')
-        plt.xlabel('RA (deg)', size=16)
-        plt.ylabel('Dec (deg)', size=16)
-        plt.title(rerunName, size=16)
-        ax2.spines['top'].set_visible(False)
-        ax2.spines['right'].set_visible(False)
-        ax2.invert_xaxis()
-        cax = plt.subplot2grid((100, 100), (60, 55), rowspan=5, colspan=45)
-        cb2 = fig.colorbar(plot2, cax=cax, orientation='horizontal', pad=0.1)
-        cb2.set_label('Number of Sources per Object', size=16)
-        cb2.set_clim(np.min(objTable.loc[dec_set2, 'nDiaSources']),
-                     np.max(objTable.loc[dec_set2, 'nDiaSources']))
-        cb2.solids.set_edgecolor("face")
-
-    else:  # one main plot
-        ax = fig.add_subplot(111)
-        plot = ax.scatter(objTable.loc[objFilter, 'ra'],
-                          objTable.loc[objFilter, 'dec'], marker='.', lw=0,
-                          s=objTable.loc[objFilter, 'nDiaSources']*8,
-                          c=objTable.loc[objFilter, 'nDiaSources'],
-                          alpha=0.7, cmap='viridis', linewidth=0.5, edgecolor='k')
-        plt.xlabel('RA (deg)', size=16)
-        plt.ylabel('Dec (deg)', size=16)
-        plt.title(rerunName, size=16)
-        ax.spines['top'].set_visible(False)
-        ax.spines['right'].set_visible(False)
-        ax.invert_xaxis()  # RA should increase to the left
-        cb = fig.colorbar(plot, orientation='horizontal')
-        cb.set_label('Number of Sources per Object', size=16)
-        cb.set_clim(np.min(objTable.loc[objFilter, 'nDiaSources']),
-                    np.max(objTable.loc[objFilter, 'nDiaSources']))
-        cb.solids.set_edgecolor("face")
+    ax = fig.add_subplot(111)
+    plot = ax.scatter(objects.loc[row_mask, 'ra'],
+                      objects.loc[row_mask, 'decl'], marker='.', lw=0,
+                      s=objects.loc[row_mask, 'nDiaSources']*8,
+                      c=objects.loc[row_mask, 'nDiaSources'],
+                      alpha=0.7, cmap='viridis', linewidth=0.5, edgecolor='k')
+    plt.xlabel('RA (deg)', size=16)
+    plt.ylabel('Dec (deg)', size=16)
+    plt.title(title, size=16)
+    ax.spines['top'].set_visible(False)
+    ax.spines['right'].set_visible(False)
+    ax.invert_xaxis()  # RA should increase to the left
+    cb = fig.colorbar(plot, orientation='horizontal')
+    cb.set_label('Number of Sources per Object', size=16)
+    cb.set_clim(np.min(objects.loc[row_mask, 'nDiaSources']),
+                np.max(objects.loc[row_mask, 'nDiaSources']))
+    cb.solids.set_edgecolor("face")
     return fig
 
 
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plot2axes(x1, y1, x2, y2, xlabel, ylabel1, ylabel2, y1err=None, y2err=None, title=''):
     """Generic plot framework for showing one y-variable on the left
     and another on the right. The x-axis is shared on the bottom.
@@ -215,6 +321,10 @@ def plot2axes(x1, y1, x2, y2, xlabel, ylabel1, ylabel2, y1err=None, y2err=None, 
     fig.tight_layout()  # otherwise the right y-label is slightly clipped
 
 
+# TODO: DM-43095, update the interface to take an instantiated butler,
+# remove pandas, and add to an appropriate visit-level (nightly?) AP pipeline.
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotDiaSourcesPerVisit(repo, sourceTable, title='',
                            instrument='DECam', collections=[]):
     """Plot DIA Sources per visit.
@@ -249,6 +359,11 @@ def plotDiaSourcesPerVisit(repo, sourceTable, title='',
               title=title)
 
 
+# TODO: DM-43095, update the interface to take an instantiated butler,
+# remove pandas, use astropy times, and add to an appropriate visit-level
+# AP pipeline.
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotDiaSourcesPerNight(sourceTable, title=''):
     """Plot DIA Sources per night.
 
@@ -284,6 +399,9 @@ def plotDiaSourcesPerNight(sourceTable, title=''):
     plt.ylim(0, np.max(visits_per_night['x'].values + 1))
 
 
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def ccd2focalPlane(x, y, ccd, camera):
     """Retrieve focal plane coordinates.
 
@@ -311,6 +429,9 @@ def ccd2focalPlane(x, y, ccd, camera):
     return point[0], point[1]
 
 
+# TODO: DM-43095, to plotUtils, without pandas
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def getCcdCorners(butler, sourceTable, instrument='DECam', collections=[]):
     """Get corner coordinates for a range of ccds.
 
@@ -363,6 +484,9 @@ def getCcdCorners(butler, sourceTable, instrument='DECam', collections=[]):
     return corners
 
 
+# TODO: DM-43095, to plotUtils
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def getCcdAndVisitSizeOnSky(repo, sourceTable, instrument='DECam',
                             collections=[], visit=None, detector=None):
     """Estimate the area of one CCD and one visit on the sky in square degrees.
@@ -408,6 +532,10 @@ def getCcdAndVisitSizeOnSky(repo, sourceTable, instrument='DECam',
     return ccdArea, visitArea
 
 
+# TODO: DM-43095, use an instantiated butler, and add to an appropriate
+# visit-level (nightly?) AP pipeline.
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotDiaSourceDensityInFocalPlane(repo, sourceTable, cmap=mpl.cm.Blues, title='',
                                      instrument='DECam', collections=[]):
     """Plot average density of DIA Sources in the focal plane (per CCD).
@@ -463,6 +591,64 @@ def plotDiaSourceDensityInFocalPlane(repo, sourceTable, cmap=mpl.cm.Blues, title
     cb.set_label('DIA Sources per sq. deg.', rotation=90)
 
 
+# TODO: DM-43095, note this plots diaObjects and is different from HitsDiaPlot,
+# yet it is HiTS-specific. For consistency, an analysis_tools style plotter
+# probably ought to live in analysis_ap, not analysis_tools.
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def plotHitsDiaObjectsOnSky(title, objects, row_mask):
+    fig = plt.figure(facecolor='white', figsize=(10, 8))
+
+    dec_set1 = (objects['decl'] > -2) & row_mask
+    dec_set2 = (~dec_set1) & row_mask
+    plt.subplots_adjust(wspace=0.1, hspace=0)
+
+    # Panel 1: one HiTS field, on the right
+    ax1 = plt.subplot2grid((100, 100), (0, 55), rowspan=90, colspan=45)
+    plot1 = ax1.scatter(objects.loc[dec_set1, 'ra'], objects.loc[dec_set1, 'decl'],
+                        marker='.', lw=0, s=objects.loc[dec_set1, 'nDiaSources']*8,
+                        c=objects.loc[dec_set1, 'nDiaSources'], alpha=0.7,
+                        cmap='viridis', linewidth=0.5, edgecolor='k')
+    plt.xlabel('RA (deg)', size=16)
+    plt.ylabel('Dec (deg)', size=16)
+    ax1.spines['top'].set_visible(False)
+    ax1.spines['left'].set_visible(False)
+    ax1.yaxis.tick_right()
+    ax1.yaxis.set_label_position('right')
+    ax1.invert_xaxis()  # RA should increase to the left
+    loc = plticker.MultipleLocator(base=0.5)  # puts ticks at regular intervals
+    ax1.yaxis.set_major_locator(loc)
+    cb = fig.colorbar(plot1, orientation='horizontal', pad=0.3)
+    cb.set_label('Number of Sources per Object', size=16)
+    cb.set_clim(np.min(objects.loc[dec_set2, 'nDiaSources']),
+                np.max(objects.loc[dec_set2, 'nDiaSources']))
+    cb.solids.set_edgecolor("face")
+    cb.remove()  # don't show colorbar by this panel
+
+    # Panel 2: two (overlapping) HiTS fields, on the left
+    ax2 = plt.subplot2grid((100, 100), (0, 0), rowspan=90, colspan=50)
+    plot2 = ax2.scatter(objects.loc[dec_set2, 'ra'], objects.loc[dec_set2, 'decl'],
+                        marker='.', lw=0, s=objects.loc[dec_set2, 'nDiaSources']*8,
+                        c=objects.loc[dec_set2, 'nDiaSources'], alpha=0.7,
+                        cmap='viridis', linewidth=0.5, edgecolor='k')
+    plt.xlabel('RA (deg)', size=16)
+    plt.ylabel('Dec (deg)', size=16)
+    plt.title(title, size=16)
+    ax2.spines['top'].set_visible(False)
+    ax2.spines['right'].set_visible(False)
+    ax2.invert_xaxis()
+    cax = plt.subplot2grid((100, 100), (60, 55), rowspan=5, colspan=45)
+    cb2 = fig.colorbar(plot2, cax=cax, orientation='horizontal', pad=0.1)
+    cb2.set_label('Number of Sources per Object', size=16)
+    cb2.set_clim(np.min(objects.loc[dec_set2, 'nDiaSources']),
+                 np.max(objects.loc[dec_set2, 'nDiaSources']))
+    cb2.solids.set_edgecolor("face")
+
+
+# TODO: as part of DM-43201, make sure the sole analysis_ap plotting
+# pipeline works, then delete this.
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotHitsSourcesOnSky(sourceTable, title=''):
     """Plot DIA Sources from three DECam HiTS fields on the sky.
 
@@ -513,6 +699,10 @@ def plotHitsSourcesOnSky(sourceTable, title=''):
     plt.subplots_adjust(wspace=0.1, hspace=0)
 
 
+# TODO: DM-43095, use analysis_tools' existing HistPlot, no hard-wired values,
+# decide how to best choose a representative seeing value for a whole visit
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotSeeingHistogram(repo, sourceTable, ccd=35, instrument='DECam', collections=[]):
     """Plot distribution of visit seeing.
 
@@ -555,6 +745,9 @@ def plotSeeingHistogram(repo, sourceTable, ccd=35, instrument='DECam', collectio
     secax.set_xlabel('Seeing FWHM (arcseconds)')
 
 
+# TODO: DM-43095, consider adding to relevant AP pipeline
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotDiaSourcesInFocalPlane(repo, sourceTable, gridsize=(400, 400), title='',
                                instrument='DECam', collections=[]):
     """Plot DIA Source locations in the focal plane.
@@ -610,13 +803,15 @@ def plotDiaSourcesInFocalPlane(repo, sourceTable, gridsize=(400, 400), title='',
     ax1.invert_xaxis()
 
 
-def plotDiaSourcesOnSkyGrid(repo, sourceTable, title=None, color='C0', size=0.1):
+# TODO: DM-43095, multiple calls to DiaSkyPlot
+# may be memory-intensive for large numbers of visits so be careful
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def plotDiaSourcesOnSkyGrid(sourceTable, title=None, color='C0', size=0.1):
     """Make a multi-panel plot of DIA Sources for each visit on the sky.
 
     Parameters
     ----------
-    repo : `str`
-        Repository corresponding to the output of an ap_pipe run.
     sourceTable : `pandas.core.frame.DataFrame`
         Pandas dataframe with DIA Sources from an APDB.
     title : `str`
@@ -648,6 +843,9 @@ def plotDiaSourcesOnSkyGrid(repo, sourceTable, title=None, color='C0', size=0.1)
         fig.suptitle(title)
 
 
+# TODO: DM-43095, and add to an appropriate AP pipeline
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotFlagHist(sourceTable, title=None,
                  badFlagList=['base_PixelFlags_flag_bad',
                               'base_PixelFlags_flag_suspect',
@@ -688,6 +886,9 @@ def plotFlagHist(sourceTable, title=None,
     plt.title(title)
 
 
+# TODO: DM-42824
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def plotFluxHistSrc(srcTable1, srcTable2=None, fluxType='psFlux',
                     label1=None, label2=None, title=None, ylog=False,
                     color1='#2979C1', color2='#Bee7F5',
@@ -738,6 +939,9 @@ def plotFluxHistSrc(srcTable1, srcTable2=None, fluxType='psFlux',
     plt.title(title)
 
 
+# TODO: DM-42824
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
 def source_magnitude_histogram(repo, sourceTable, bandToPlot, instrument,
                                collections, detectorToUse=42, binsize=0.2,
                                min_magnitude=None, max_magnitude=None,
@@ -817,96 +1021,241 @@ def source_magnitude_histogram(repo, sourceTable, bandToPlot, instrument,
     ax.set_yscale('log')
 
 
-# Stuff from plotLightCurve.py
-
-def retrieveCutouts(butler, dataId, collections, center, size=lsst.geom.Extent2I(30, 30), diffName='deep'):
-    """Return small cutout exposures for a science exposure, difference image,
-    and warped template.
-
-    Parameters
-    ----------
-    butler : `lsst.daf.butler.Butler`
-        Butler in the repository corresponding to the output of an ap_pipe run.
-    dataId : `dict`-like
-        Gen3 data ID specifying at least instrument, visit, and detector.
-    collections : `str` or `list`
-        Gen3 collection or collections from which to load the exposures.
-    center : `lsst.geom.SpherePoint`
-        Desired center coordinate of cutout.
-    size : `lsst.geom.Extent`, optional
-        Desired size of cutout, default is 30x30 pixels
-    diffName : `str`, optional
-        Default is 'deep', but 'goodSeeing' may be needed instead.
-
-    Returns
-    -------
-    scienceCutout : `lsst.afw.Exposure`
-        Cutout of calexp at location 'center' of size 'size'.
-    differenceCutout : `lsst.afw.Exposure`
-        Cutout of diffName_differenceExp at location 'center' of size 'size'.
-    templateCutout : `lsst.afw.Exposure`
-        Cutout of diffName_templateExp at location 'center' of size 'size'.
-    """
-    science, difference, template = loadExposures(butler, dataId,
-                                                  collections, diffName)
-    scienceCutout = science.getCutout(center, size)
-    differenceCutout = difference.getCutout(center, size)
-    templateCutout = template.getCutout(center, size)
-    return scienceCutout, differenceCutout, templateCutout
-
-
-def plotCutout(scienceCutout, differenceCutout, templateCutout, output=None):
-    """Plot the cutouts for one DIASource in one image.
+# TODO: DM-40203
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def plotTractOutline(tractInfo, patchList, axes=None, fontSize=10,
+                     maxDegBeyondPatch=1.5):
+    """Plot the the outline of the tract and patches highlighting
+    those with data.
+    As some skyMap settings can define tracts with a large number of patches,
+    this can become very crowded.
+    So, if only a subset of patches are included, find the outer boundary
+    of all patches in patchList and only plot to maxDegBeyondPatch degrees
+    beyond those boundaries (in all four directions).
 
     Parameters
     ----------
-    scienceCutout : `lsst.afw.Exposure`
-        Cutout of calexp returned by retrieveCutouts.
-    differenceCutout : `lsst.afw.Exposure`
-        Cutout of deepDiff_differenceExp returned by retrieveCutouts.
-    templateCutout : `lsst.afw.Exposure`
-        Cutout of deepDiff_templateExp returned by retrieveCutouts.
-    output : `str`, optional
-        If provided, save png to disk at output filepath.
+    tractInfo : `lsst.skymap.tractInfo.ExplicitTractInfo`
+       Tract information object for extracting tract RA and DEC limits.
+    patchList : `list` of `str`
+       List of patch IDs with data to be plotted.  These will be color
+       shaded in the outline plot.
+    fontSize : `int`
+       Font size for plot labels.
+    maxDegBeyondPatch : `float`
+       Maximum number of degrees to plot beyond the border defined by all
+       patches with data to be plotted.
     """
-    def do_one(ax, data, name):
-        interval = aviz.ZScaleInterval()
-        if name == 'Difference':
-            norm = aviz.ImageNormalize(data, stretch=aviz.LinearStretch())
-        else:
-            norm = aviz.ImageNormalize(data, interval=interval, stretch=aviz.AsinhStretch(a=0.01))
-        ax.imshow(data, cmap=cm.bone, interpolation="none", norm=norm)
-        ax.axis('off')
-        ax.set_title(name)
+    def percent(values, p=0.5):
+        """Return a value a fraction of the way between the min and max
+        values in a list.
+        """
+        interval = max(values) - min(values)
+        return min(values) + p*interval
 
-    fig, (ax1, ax2, ax3) = plt.subplots(1, 3)
-    do_one(ax1, templateCutout.image.array, "Template")
-    do_one(ax2, scienceCutout.image.array, "Science")
-    do_one(ax3, differenceCutout.image.array, "Difference")
-    plt.tight_layout()
+    if axes is None:
+        fig = plt.figure(figsize=(8, 8))
+        axes = fig.gca()
+    buff = 0.02
+    axes.tick_params(which="both", direction="in", labelsize=fontSize)
+    axes.locator_params(nbins=6)
+    axes.ticklabel_format(useOffset=False)
 
-    if output is not None:
-        plt.savefig(output, bbox_inches="tight")
-        plt.close()
+    tractRa, tractDec = bboxToRaDec(tractInfo.getBBox(), tractInfo.getWcs())
+    patchBoundary = getRaDecMinMaxPatchList(patchList, tractInfo,
+                                            pad=maxDegBeyondPatch)
+
+    xMin = min(max(tractRa), patchBoundary.raMax) + buff
+    xMax = max(min(tractRa), patchBoundary.raMin) - buff
+    yMin = max(min(tractDec), patchBoundary.decMin) - buff
+    yMax = min(max(tractDec), patchBoundary.decMax) + buff
+    xlim = xMin, xMax
+    ylim = yMin, yMax
+    axes.fill(tractRa, tractDec, fill=True, color="black", lw=1, linestyle='solid',
+              edgecolor='k', alpha=0.2)
+    for ip, patch in enumerate(tractInfo):
+        patchIndexStr = str(patch.getIndex()) + ' ' + str(tractInfo.getSequentialPatchIndex(patch))
+        patchWithData = tractInfo.getSequentialPatchIndex(patch)
+        color = "k"
+        alpha = 0.05
+        if patchWithData in patchList:
+            # color = ("c", "g", "r", "b", "m")[ip%5]
+            color = 'g'
+            alpha = 0.5
+        ra, dec = bboxToRaDec(patch.getOuterBBox(), tractInfo.getWcs())
+        deltaRa = abs(max(ra) - min(ra))
+        deltaDec = abs(max(dec) - min(dec))
+        pBuff = 0.5*max(deltaRa, deltaDec)
+        centerRa = min(ra) + 0.5*deltaRa
+        centerDec = min(dec) + 0.5*deltaDec
+        if (centerRa < xMin + pBuff and centerRa > xMax - pBuff
+                and centerDec > yMin - pBuff and centerDec < yMax + pBuff):
+            axes.fill(ra, dec, fill=True, color=color, lw=1,
+                      linestyle="solid", alpha=alpha)
+            if patchWithData in patchList or (centerRa < xMin - 0.2*pBuff
+                                              and centerRa > xMax + 0.2*pBuff
+                                              and centerDec > yMin + 0.2*pBuff
+                                              and centerDec < yMax - 0.2*pBuff):
+                axes.text(percent(ra), percent(dec, 0.5), str(patchIndexStr),
+                          fontsize=fontSize - 1, horizontalalignment="center",
+                          verticalalignment="center")
+    axes.set_xlabel('RA', size='large')
+    axes.set_ylabel('Dec', size='large')
+    axes.set_xlim(xlim)
+    axes.set_ylim(ylim)
 
 
-def setObjectFilter(objTable):
-    """Define a subset of objects to plot, i.e., make some kind of quality cut.
-
-    The definition of objFilter is presently hard-wired here.
+# TODO: DM-40203
+@deprecated(reason="This function is deprecated and will be removed soon.",
+            version="v27", category=FutureWarning)
+def mosaicCoadd(repo, patch_list, band='g', tractIndex=0, refPatchIndex=None,
+                sampling=100, norm=None, nImage=False, fig=None,
+                show_colorbar=True, filename=None, flipX=True, flipY=False,
+                instrument=None, collections=None, skymapName=None,
+                coaddName='deepCoadd'):
+    """Generate a mosaic image of many coadded patches. Gen3 only.
 
     Parameters
     ----------
-    objTable : `pandas.DataFrame`
-        DIA Object Table.
-
-    Returns
-    -------
-    objFilter : `pandas.Series` of `bool`
-        Filter applied to create a subset (e.g., quality cut) from objTable.
+    repo : `str`
+        The path to the data repository.
+    patch_list : `list`
+        A list of the patch indices containing images to mosaic.
+        List elements will be `float` (sequential patch indices).
+    band : `str`, optional
+        The band of the coadd to retrieve from the repository.
+    tractIndex : `int`, optional
+        The tract of the skyMap.
+    refPatchIndex : `str`, optional
+        If set, use the given patch to set the image normalization
+        for the figure.
+    sampling : `int`, optional
+        Stride factor to sample each input image in order to reduce the
+        size in memory.
+        A `sampling` of 1 will attempt to display every pixel.
+    norm : `astropy.visualization.ImageNormalize`, optional
+        Normalization to set the color scale of the images.
+        If `None`, the normalization will be calculated from the first image.
+        If you wish to use any normalization other than zscale, you must
+        calculate it ahead of time and pass it in as `norm` here.
+    nImage : `bool`, optional
+        Mosaic the nImage instead of the coadd.
+    fig : `matplotlib.pyplot.fig`, optional
+        Figure instance to display the mosaic in.
+    show_colorbar : `bool`, optional
+        Display a colorbar on the figure.
+    filename : `str`, optional
+        If set, write the figure to a file with the given filename.
+    flipX : `bool`, optional
+        Set to flip the individual patch images horizontally.
+    flipY : `bool`, optional
+        Set to flip the individual patch images vertically.
+    instrument : `str`, optional
+        Default is 'DECam'.
+    collections : `list` or `str`, optional
+        Must be provided to load the camera properly
+    skymapName : `str`, optional
+        Must be provided to load the skymap. e.g., 'hsc_rings_v1'
+    coaddName : `str`, optional
+        Type of coadd, default `deepCoadd` (you might want `goodSeeingCoadd`)
     """
-    objFilter = ((objTable['nDiaSources'] > 14) & (objTable['flags'] == 0))
-    numTotal = len(objTable['diaObjectId'])
-    numFilter = len(objTable.loc[objFilter, 'diaObjectId'])
-    print('There are {0} total DIAObjects and {1} filtered DIAObjects.'.format(numTotal, numFilter))
-    return objFilter
+    # Decide what data product to plot (actual pixel data or nImages)
+    if nImage:
+        coaddName += '_nImage'
+
+    # Create a mapping between sequential patch indices and (x,y) patch indices
+    if not collections:
+        raise ValueError('One or more collections is required.')
+    if not skymapName:
+        raise ValueError('A skymap is required.')
+    if not instrument:
+        raise ValueError('An instrument is required.')
+    butler = dafButler.Butler(repo, collections=collections)
+    skymap = butler.get('skyMap', skymap=skymapName, instrument=instrument,
+                        collections='skymaps')
+    tract = skymap.generateTract(tractIndex)
+    patchesToLoad = [patch for patch in tract if tract.getSequentialPatchIndex(patch) in patch_list]
+    patchIndicesToLoad = [patch.getIndex() for patch in patchesToLoad]  # (x,y) indices
+    indexmap = {}
+    for patch in tract:
+        indexmap[str(tract.getSequentialPatchIndex(patch))] = patch.getIndex()
+
+    # Use a reference patch to set the normalization for all the patches
+    if norm is None:
+        if refPatchIndex in patchIndicesToLoad:
+            sequentialPatchIndex = [key for key, value in indexmap.items() if value == refPatchIndex][0]
+            dataId = {'band': band, 'tract': tractIndex,
+                      'patch': int(sequentialPatchIndex),
+                      'instrument': instrument, 'skymap': skymapName}
+            if nImage:
+                coaddArray = butler.get(coaddName, dataId=dataId).getArray()
+            else:
+                coaddArray = butler.get(coaddName, dataId=dataId).getImage().getArray()
+            norm = ImageNormalize(coaddArray, interval=ZScaleInterval(),
+                                  stretch=SqrtStretch())
+
+    # Set up the figure grid
+    patch_x = []
+    patch_y = []
+    for patch in patchesToLoad:
+        patch_x.append(patch.getIndex()[0])
+        patch_y.append(patch.getIndex()[1])
+    x0 = min(patch_x)
+    y0 = min(patch_y)
+    nx = max(patch_x) - x0 + 1
+    ny = max(patch_y) - y0 + 1
+    fig = plt.figure(figsize=(nx, ny), constrained_layout=False)
+    gs1 = fig.add_gridspec(ny, nx, wspace=0, hspace=0)
+
+    # Plot coadd patches with data and print the patch index if there's no data
+    for x in range(nx):
+        for y in range(ny):
+            figIdx = x + nx*y
+            ax = fig.add_subplot(gs1[figIdx])
+            ax.set_aspect('equal', 'box')
+            patchIndex = (x, ny - y - 1)
+            if patchIndex in patchIndicesToLoad:
+                sequentialPatchIndex = [key for key, value in indexmap.items() if value == patchIndex][0]
+                dataId = {'band': band, 'tract': tractIndex, 'patch': int(sequentialPatchIndex),
+                          'instrument': instrument, 'skymap': skymapName}
+                try:
+                    coadd = butler.get(coaddName, dataId=dataId)
+                except LookupError:
+                    print(f'Failed to retrieve data for patch {patchIndex}')
+                    ax.text(.3, .3, patchIndex)
+                    continue
+                if nImage:
+                    coaddArray = coadd.getArray()
+                else:
+                    coaddArray = coadd.getImage().getArray()
+                coaddArray = coaddArray[::sampling, ::sampling]
+                if flipX:
+                    coaddArray = np.flip(coaddArray, axis=0)
+                if flipY:
+                    coaddArray = np.flip(coaddArray, axis=1)
+                if norm is None:
+                    norm = ImageNormalize(coaddArray,
+                                          interval=ZScaleInterval(),
+                                          stretch=SqrtStretch())
+                im = ax.imshow(coaddArray, cmap='gray', norm=norm)
+            else:
+                ax.text(.3, .3, patchIndex)
+                im = None
+                # print('No data in patch', patchIndex)
+            plt.setp(ax, xticks=[], yticks=[])
+
+    # Adjust and annotate plot
+    if show_colorbar and (im is not None):
+        cbar_width = 0.01
+        cbar_height = 0.5
+        cbar_ax = fig.add_axes([0.9 - cbar_width, 0.5 - cbar_height/2,
+                                cbar_width, cbar_height])
+        fig.colorbar(im, cax=cbar_ax)
+
+    # Save figure, if desired
+    if filename:
+        try:
+            plt.savefig(filename, transparent=True)
+        except Exception as e:
+            print(f"Could not write file '{filename}': {e}")
